OSv Init: http client
The http client is loosely based on the boost http client example.
The original example can be found here:
http://www.boost.org/doc/libs/1_55_0/doc/html/boost_asio/example/cpp03/http/client/sync_client.cpp

The client class redirect its output to an output stream so if the c is
a client the following can be done:
std::cout << c;

The client is used to retrieve init files from a remote server. It can
be replaced by a library client if need arises.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>