OSv Init: the osvinit class
The osvinit object uses command from a yaml parser to either call an API
command or to perform internal command.
The supported options are:
Each command contains: GET, PUT, POST or DELETE with a url of the
command. Path parameters comes as part of the url.
Query parameteres follow as a yaml parameters.

The osvinit can be set to stop on errors or ignore them.

When opening the api, the osv init creates a new thread for that, it
would wait for that thread to terminate.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>