Http Server: A singleton class for starting the httpserver
The osvinit needs access to the httpserver internal logic before it
starts to serve external clients. It also needs the ability to delay its
operation until the initilization process completed.

global server is a singleton class that holds an httpserver and
responsible for its initialization.

To access the httpserver, it expose the server routes object that can be
used directly by the OSv Init.
The server itself needs to be started explicitely to start serving
external clients.

The logic in the global_server replaces the code that used to be in
main.

Instead of starting an httpserver by its own, main now calls the
global_server to run the server.

The route initilization is done by the global_server class.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>